+++ packages/analytics/analytics-chart/src/components/CsvExportModal.cy.ts
@@ -1,6 +1,6 @@
 // Cypress component test spec file
 import CsvExportModal from './CsvExportModal.vue'
-import { exploreResult, exploreV2Result, emptyExploreResult } from '../../fixtures/mockData'
+import { exploreResult, emptyExploreResult } from '../../fixtures/mockData'
 
 const DOWNLOADS_FOLDER = Cypress.config('downloadsFolder')
 const MAX_ROWS = 3
@@ -23,7 +23,7 @@ describe('<CsvExportModal />', () => {
     cy.getTestId('csv-download-button').should('be.disabled')
   })
 
-  it('Export Modal with v1 explore data', () => {
+  it('Export Modal with explore data', () => {
     cy.mount(CsvExportModal, {
       props: {
         chartData: exploreResult,
@@ -49,30 +49,15 @@ describe('<CsvExportModal />', () => {
     cy.getTestId('csv-export-modal').find('.k-table thead th').should(th => {
       const elements = Array.from(th, e => e.innerText)
 
-      expect(elements).eql(['Timestamp', 'UtcOffset', 'StatusCode', 'TotalRequests'])
+      expect(JSON.stringify(elements)).to.equal('["Timestamp","UTC Offset","Status Code","Request Count"]')
     })
 
     // Save to CSV and check actual contents
     cy.getTestId('csv-download-button').click()
 
     cy.readFile(`${DOWNLOADS_FOLDER}/total-requests-${new Date().toISOString().slice(0, 10)}.csv`).then(contents => {
-      expect(contents).contain('Timestamp,UtcOffset,StatusCode,TotalRequests')
-      expect(contents).contain(',300,1239')
+      expect(contents).contain('Timestamp,UTC Offset,Status Code,Request Count')
+      expect(contents).contain('500,10022')
     })
-  })
-
-  it('Export Modal with v2 explore data', () => {
-    cy.mount(CsvExportModal, {
-      props: {
-        filename: 'Total requests',
-        chartData: exploreV2Result,
-      },
-    })
-
-    cy.getTestId('csv-export-modal').should('exist')
-    cy.get('.k-table-empty-state').should('not.exist')
-    cy.get('.modal-content .vitals-table').should('exist')
-    cy.get('.modal-content .vitals-table').should('exist')
-    cy.getTestId('csv-download-button').should('not.be.disabled')
   })
 })