+++ packages/analytics/analytics-utilities/src/types/explore-v4.ts
@@ -90,11 +90,21 @@ export interface RelativeTimeRangeV4 {
 
 export type TimeRangeV4 = AbsoluteTimeRangeV4 | RelativeTimeRangeV4
 
+export const granularityValues = [
+  'minutely',
+  'hourly',
+  'daily',
+  'weekly',
+  'trend',
+] as const
+
+export type GranularityValues = typeof granularityValues[number]
+
 export interface ExploreQuery {
   metrics?: ExploreAggregations[]
   dimensions?: QueryableExploreDimensions[]
   filters?: ExploreFilter[]
-  granularity_ms?: number
+  granularity?: GranularityValues
   time_range: TimeRangeV4
   limit?: number
   descending?: boolean
@@ -122,8 +132,7 @@ export interface QueryResponseMeta {
   display: DisplayBlob
   metric_names?: ExploreAggregations[]
   metric_units?: MetricUnit
-  granularity?: number
-  granularity_ms?: number
+  granularity_ms: number
   truncated?: boolean
   limit?: number
   query_id: string