+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.3.15...@kong-ui-public/entities-upstreams-targets@2.3.16) (2024-02-07)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 ## [2.3.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.3.14...@kong-ui-public/entities-upstreams-targets@2.3.15) (2024-02-05)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets